requirements.txt
annotated-types==0.7.0
bitarray==3.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
ckzg==2.0.1
cytoolz==1.0.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
hexbytes==1.2.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
parsimonious==0.10.0
pluggy==1.5.0
poly_eip712_structs==0.0.1
py-clob-client==0.18.0
py_order_utils==0.3.2
pycryptodome==3.21.0
pydantic==2.10.2
pydantic_core==2.27.1
pytest==8.3.4
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
toolz==1.0.0
typing_extensions==4.12.2
urllib3==2.2.3
